Refactored Chart and extracted common functions to mixin
