+++ resources/js/components/ChartjsLineChart.vue
@@ -1,11 +1,13 @@
 <script>
     import { Line, mixins } from 'vue-chartjs'
+    import charts from "../mixins/charts";
+
     const {reactiveData} = mixins;
 
     export default {
         extends: Line,
 
-        mixins: [reactiveData],
+        mixins: [reactiveData, charts],
 
         props: {
             'dataset': Array,
@@ -18,26 +20,6 @@
             //we need to set options manually as options are used to re-render chart when data changes in reactiveData/reactiveProp mixin
             this.options = this.settings.options;
             this.renderChart(this.chartData, this.options);
-        },
-
-        methods: {
-            createChartDataset: function(){
-                let datasets = [...this.additionalDatasets];
-
-                for (let data in this.dataset) {
-                    datasets.unshift(
-                        {
-                            ...this.dataset[data],
-                            ...{fill:false}
-                        }
-                    );
-                }
-
-                return {
-                    labels: this.settings.parameters,
-                    datasets: datasets
-                }
-            }
         },
 
         watch: {
